FlaskApp/requirements.txt
requests==2.22.0
Flask==2.0.1
pytest==6.2.4
python-dotenv==0.19.1
